payload too large working

+++ srcs/config/LocationConfig.cpp
@@ -6,7 +6,7 @@
 /*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/07 23:05:41 by jdagoy            #+#    #+#             */
-/*   Updated: 2024/10/22 15:06:35 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/10/23 00:16:12 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -164,13 +164,13 @@ std::string LocationConfig::getRedirect() const
     return (std::string());
 }
 
-int LocationConfig::getClientMaxBodySize()
+size_t LocationConfig::getClientMaxBodySize()
 {
     std::map<std::string, std::string>::const_iterator directive;
     for (directive = _directives.begin(); directive != _directives.end(); directive++)
     {
         if (directive->first == "client_max_body_size")
-            return (strToInt(directive->second));
+            return (static_cast<size_t>(strToInt(directive->second)));
     }
     return (-1);
 }
